feat: 测测测测

+++ build/ci.js
@@ -3,13 +3,17 @@ const ci = require('miniprogram-ci')
 const fs = require('fs')
 const packageJson = require('../package.json')
 
-const privateKeyContent = process.env.WX_PRIVATE_KEY
-if (!privateKeyContent) {
-  throw new Error('未找到私钥内容，请确保已正确配置 GitHub Secrets')
-}
-
 const privateKeyPath = path.resolve(__dirname, './key')
-fs.writeFileSync(privateKeyPath, privateKeyContent)
+
+// 检查私钥文件是否已存在
+if (!fs.existsSync(privateKeyPath)) {
+  const privateKeyContent = process.env.WX_PRIVATE_KEY
+  if (!privateKeyContent) {
+    throw new Error('未找到私钥内容，请确保已正确配置 GitHub Secrets')
+  }
+  console.log('>>>>写入私钥文件：', privateKeyPath);
+  fs.writeFileSync(privateKeyPath, privateKeyContent)
+}
 
 const project = new ci.Project({
   appid: 'wx622bee4f78fa4f5a',
